clippy + readme format

+++ crates/bevy-inspector-egui/examples/integrations/egui_dock.rs
@@ -108,8 +108,8 @@ fn set_camera_viewport(
 
     let scale_factor = window.scale_factor() * egui_settings.scale_factor;
 
-    let viewport_pos = ui_state.viewport_rect.left_top().to_vec2() * scale_factor as f32;
-    let viewport_size = ui_state.viewport_rect.size() * scale_factor as f32;
+    let viewport_pos = ui_state.viewport_rect.left_top().to_vec2() * scale_factor;
+    let viewport_size = ui_state.viewport_rect.size() * scale_factor;
 
     cam.viewport = Some(Viewport {
         physical_position: UVec2::new(viewport_pos.x as u32, viewport_pos.y as u32),
